Big Data analysis for Airline delays with Revolution R
In this course, I analyzed and predicted the delays based on day of week
using R for data analysis and mini-chart plots using R libraries. I
developed model using rxLinModel, use of R functions like rxSummary,
rxQuantile, rxCrossTabs, rxCube to analyze Aircraft delays and the
relationships with the week day, holiday and any deviations from
standard schedules. I applied transformations for dataframe to XDF
format and used Transforms/ TransformFunc for data manipulation and
display using Histogram to verify the deviations. Use this model to
predict the possibility of delay of the next flight. Successfully
completed the assessment with a certificate